Finished restructuring services directory. Added first service unit test.
