+++ .gitignore
@@ -50,3 +50,6 @@ env.yml
 
 # Jetbrains IDEs
 .idea
+
+# Visual Studio Code
+.vscode
